+++ src/main/java/pm/axe/internal/HasTabInit.java
@@ -1,0 +1,7 @@
+package pm.axe.internal;
+
+import pm.axe.db.models.User;
+
+public interface HasTabInit {
+    void tabInit(User user);
+}
